adds labels and private variables to badge type file
